refactor(tangle-dapp): Move `bridge` types to `tangle-shared-ui`
