+++ apps/tangle-dapp/context/BridgeContext.tsx
@@ -6,6 +6,7 @@ import { chainsConfig } from '@webb-tools/dapp-config/chains/chain-config';
 import { ChainConfig } from '@webb-tools/dapp-config/chains/chain-config.interface';
 import getChainFromConfig from '@webb-tools/dapp-config/utils/getChainFromConfig';
 import { calculateTypedChainId } from '@webb-tools/sdk-core/typed-chain-id';
+import { BridgeType } from '@webb-tools/tangle-shared-ui/types/bridge';
 import assert from 'assert';
 import {
   createContext,
@@ -23,7 +24,6 @@ import {
   BridgeFeeItem,
   BridgeFeeType,
   BridgeTokenId,
-  BridgeType,
   BridgeWalletError,
 } from '../types/bridge';
 import { isEVMChain } from '../utils/bridge';
